Add functionality to map facts to column names.
Supply a string to output_as_csv() that maps fact names to column names.
Column name is optional; if not supplied the fact name will be used for
the column name.

e.g. "certname=hostname,operatingsystem=operating system,operatingsystemmajrelease"

This will create a CSV file with header line:

hostname,operating system,operatingsystemmajrelease

Note that if this option is used no other facts in the source JSON file
will appear in the CSV output.