Add mail_report.sh and sample.report.
mail_report.sh is eventually intended to be run from cron, getting its
settings from a ".report" file modelled on sample.report.

With different ".report" files it should be possible to generate
different reports (eg for use with iTop.)  This may be a useful
replacement for MCollective, or it may not.