Allow the use of stdin as input.
This means that the file argument can now either be a string (file name)
or a file object (stdin).  As a result, load_json_data() has been
refactored to remove the file-handling aspects.  This code now resides
in the main() function.  This was done mainly because I don't like
having function arguments that can be multiple types.